requirements.txt
# Python pakages required for main.py

ffmpeg-python
